refactor: Move OsmTagMapperSource out of class
