+++ src/ext/java/org/opentripplanner/ext/reportapi/resource/ReportResource.java
@@ -17,7 +17,7 @@ import org.opentripplanner.ext.reportapi.model.GraphReportBuilder;
 import org.opentripplanner.ext.reportapi.model.GraphReportBuilder.GraphStats;
 import org.opentripplanner.ext.reportapi.model.TransfersReport;
 import org.opentripplanner.model.transfer.TransferService;
-import org.opentripplanner.openstreetmap.tagmapping.OsmTagMapper;
+import org.opentripplanner.openstreetmap.tagmapping.OsmTagMapperSource;
 import org.opentripplanner.standalone.api.OtpServerRequestContext;
 import org.opentripplanner.transit.service.TransitService;
 
@@ -63,7 +63,7 @@ public class ReportResource {
   public Response getBicycleSafetyAsCsv(
     @DefaultValue("default") @QueryParam("osmWayPropertySet") String osmWayPropertySet
   ) {
-    var source = OsmTagMapper.Source.valueOf(osmWayPropertySet);
+    var source = OsmTagMapperSource.valueOf(osmWayPropertySet);
     return Response
       .ok(BicycleSafetyReport.makeCsv(source))
       .header(
