+++ src/main/java/org/opentripplanner/graph_builder/module/osm/parameters/OsmExtractParameters.java
@@ -4,16 +4,16 @@ import java.net.URI;
 import java.time.ZoneId;
 import javax.annotation.Nullable;
 import org.opentripplanner.graph_builder.model.DataSourceConfig;
-import org.opentripplanner.openstreetmap.tagmapping.OsmTagMapper;
+import org.opentripplanner.openstreetmap.tagmapping.OsmTagMapperSource;
 
 /**
  * Configure an OpenStreetMap extract.
  * Example: {@code "osm" : [ {source: "file:///path/to/otp/norway.pbf"} ] }
  *
  */
-public record OsmExtractParameters(URI source, OsmTagMapper.Source osmTagMapper, ZoneId timeZone)
+public record OsmExtractParameters(URI source, OsmTagMapperSource osmTagMapper, ZoneId timeZone)
   implements DataSourceConfig {
-  public static final OsmTagMapper.Source DEFAULT_OSM_TAG_MAPPER = OsmTagMapper.Source.DEFAULT;
+  public static final OsmTagMapperSource DEFAULT_OSM_TAG_MAPPER = OsmTagMapperSource.DEFAULT;
 
   public static final ZoneId DEFAULT_TIME_ZONE = null;
