+++ src/main/java/org/opentripplanner/graph_builder/module/osm/parameters/OsmExtractParametersBuilder.java
@@ -2,7 +2,7 @@ package org.opentripplanner.graph_builder.module.osm.parameters;
 
 import java.net.URI;
 import java.time.ZoneId;
-import org.opentripplanner.openstreetmap.tagmapping.OsmTagMapper;
+import org.opentripplanner.openstreetmap.tagmapping.OsmTagMapperSource;
 
 /**
  * Configure an OpenStreetMap extract.
@@ -17,7 +17,7 @@ public class OsmExtractParametersBuilder {
   /**
    * Custom OSM way properties for this extract.
    */
-  private OsmTagMapper.Source osmTagMapper;
+  private OsmTagMapperSource osmTagMapper;
 
   /**
    * The timezone to use to resolve opening hours in this extract.
@@ -39,7 +39,7 @@ public class OsmExtractParametersBuilder {
     return this;
   }
 
-  public OsmExtractParametersBuilder withOsmTagMapper(OsmTagMapper.Source mapper) {
+  public OsmExtractParametersBuilder withOsmTagMapper(OsmTagMapperSource mapper) {
     this.osmTagMapper = mapper;
     return this;
   }
@@ -53,7 +53,7 @@ public class OsmExtractParametersBuilder {
     return source;
   }
 
-  public OsmTagMapper.Source getOsmTagMapper() {
+  public OsmTagMapperSource getOsmTagMapper() {
     return osmTagMapper;
   }
